Combined RBT_SET_GAIN and RBT_SET_OFFSET commands in the single command RBT_SET_POT as both commands simple set the specified digital pot.
Added code to read the serial port register to clear the interrupt when
a byte is received when the SERIAL_PACKET_READY flag is still set (has
not yet been processed by the main loop). Not reading the byte in that case
caused a continuous interrupt loop. Need to add code to host to prevent
sending packets too fast.

Signed-off-by: Mike Schoonover <mike@skoonie.com>