Xmt of Monitor Packet after Monitor Mode started works!!!!
We were overwriting whatever was in the xmt buffer before it had been
transmitted. We never had this issue before because he was never
autonomously sending back packets; they only sent when requested. We
never received a second command for information while still processing the
first command. So, each time the serial output buffer was written to, it
had already been sent to and handled by the Rabbit.

To solve this, I put a check at the top of process monitor to see if the
transmit interrupt was enabled. If it is, then data is still being sent
and the function bails without doing anything.